Move flutter file under flutter/ directory
